+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to the **FIX Master** extension will be documented in this f
 
 This format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## [unreleased] 
+## [2.2.0] 2025-12-24 
 
 ### Added
+
 - Command `FIX Master - Convert to Raw FIX` which will convert any pretty printed FIX messages in the current file to raw FIX. 
 - Command `FIX Master - Convert Selection to Raw FIX` which will convert any pretty printed FIX messages in the current selection to raw FIX.